Поправила Code style + добавила модификатор доступа
