+++ src/main/java/pojo/OrderBody.java
@@ -1,4 +1,4 @@
-package POJO;
+package pojo;
 
 public class OrderBody {
         String[] ingredients;
