+++ src/main/java/pojo/UserData.java
@@ -1,4 +1,4 @@
-package POJO;
+package pojo;
 
 public class UserData {
