+++ src/test/java/apitests/CreatingUserTest.java
@@ -1,7 +1,7 @@
-package ApiTests;
+package apitests;
 
-import POJO.LoginData;
-import POJO.UserData;
+import pojo.LoginData;
+import pojo.UserData;
 import io.qameta.allure.Description;
 import io.qameta.allure.junit4.DisplayName;
 import io.restassured.RestAssured;
@@ -9,12 +9,11 @@ import io.restassured.response.Response;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
-
 import static io.restassured.RestAssured.given;
 import static org.hamcrest.CoreMatchers.equalTo;
 
 public class CreatingUserTest {
-    String token;
+    private String token;
 
     @Before
     public void setUp() {
