+++ src/test/java/apitests/GetOrderTest.java
@@ -1,21 +1,19 @@
-package ApiTests;
+package apitests;
 
-import POJO.LoginData;
-import POJO.UserData;
-import org.junit.Test;
+import pojo.LoginData;
+import pojo.UserData;
 import io.qameta.allure.Description;
 import io.qameta.allure.junit4.DisplayName;
+import org.junit.Test;
 import io.restassured.RestAssured;
 import io.restassured.response.Response;
 import org.junit.After;
 import org.junit.Before;
-
 import static io.restassured.RestAssured.given;
 import static org.hamcrest.CoreMatchers.equalTo;
 
-public class ChangingUserTest {
-
-    String token;
+public class GetOrderTest {
+    private String token;
 
     @Before
     public void setUp() {
@@ -23,36 +21,30 @@ public class ChangingUserTest {
     }
 
     @Test
-    @DisplayName("Changing users data - authorized")
-    @Description("Changing users data - authorized attempt")
-    public void ChangingUserDataAuthorizedTest(){
+    @DisplayName("Get order authorized")
+    @Description("Getting order while being authorized")
+    public void GetOrdersAuthorized(){
         UserData testUser = new UserData("Tapioka" + Math.random()*1000 + "@ya.ru", "qwerty" + Math.random()*1000,
                 "HungryUser" + Math.random()*1000);
         Response testResponse = given().header("Content-Type","application/json").body(testUser)
                 .post("/api/auth/register");
         token = testResponse.body().as(LoginData.class).getAccessToken().substring(7);
-        UserData testUser2 = new UserData("Tapioka" + Math.random()*1000 + "@ya.ru", "qwerty" + Math.random()*1000,
-                "HungryUser" + Math.random()*1000);
         given().header("Content-Type","application/json").auth()
-                .oauth2(token).body(testUser2).patch("/api/auth/user")
-                .then().statusCode(200).and()
+                .oauth2(token).get("/api/orders").then().statusCode(200).and()
                 .assertThat().body("success", equalTo(true));
     }
 
     @Test
-    @DisplayName("Changing users data - not authorized")
-    @Description("Changing users data - not authorized attempt")
-    public void ChangingUserDataNotAuthorizedTest(){
+    @DisplayName("Get order without authorization")
+    @Description("Trying to get order without authorization")
+    public void GetOrdersNotAuthorized(){
         UserData testUser = new UserData("Tapioka" + Math.random()*1000 + "@ya.ru", "qwerty" + Math.random()*1000,
                 "HungryUser" + Math.random()*1000);
         Response testResponse = given().header("Content-Type","application/json").body(testUser)
                 .post("/api/auth/register");
         token = testResponse.body().as(LoginData.class).getAccessToken().substring(7);
-        UserData testUser2 = new UserData("Tapioka" + Math.random()*1000 + "@ya.ru", "qwerty" + Math.random()*1000,
-                "HungryUser" + Math.random()*1000);
-        given().header("Content-Type","application/json").body(testUser2).patch("/api/auth/user")
-                .then().statusCode(401).and()
-                .assertThat().body("success", equalTo(false));
+        given().header("Content-Type","application/json").get("/api/orders").then().statusCode(401)
+            .and().assertThat().body("message", equalTo("You should be authorised"));
     }
 
     @After
